feat(rust): rename dynvault -> vault
