+++ implementations/rust/channel/src/lib.rs
@@ -35,7 +35,7 @@ use ockam_message::message::{
 use ockam_system::commands::OckamCommand::Router;
 use ockam_system::commands::{ChannelCommand, OckamCommand, RouterCommand};
 use ockam_vault::types::PublicKey;
-use ockam_vault::{DynVault, Secret};
+use ockam_vault::{Secret, Vault};
 use rand::{thread_rng, Rng};
 use std::{
     collections::BTreeMap,
@@ -65,7 +65,7 @@ pub struct ChannelManager<
     rx: Receiver<OckamCommand>,
     tx: Sender<OckamCommand>,
     router_tx: Sender<OckamCommand>,
-    vault: Arc<Mutex<dyn DynVault + Send>>,
+    vault: Arc<Mutex<dyn Vault + Send>>,
     new_key_exchanger: E,
     phantom_i: PhantomData<I>,
     phantom_r: PhantomData<R>,
@@ -91,7 +91,7 @@ impl<I: KeyExchanger, R: KeyExchanger, E: NewKeyExchanger<I, R>> ChannelManager<
         rx: Receiver<OckamCommand>,
         tx: Sender<OckamCommand>,
         router_tx: Sender<OckamCommand>,
-        vault: Arc<Mutex<dyn DynVault + Send>>,
+        vault: Arc<Mutex<dyn Vault + Send>>,
         new_key_exchanger: E,
         resp_key_ctx: Option<Arc<Box<dyn Secret>>>,
         init_key_ctx: Option<Arc<Box<dyn Secret>>>,
