+++ implementations/rust/vault/src/lib.rs
@@ -251,127 +251,3 @@ pub trait DynVault {
         data: &[u8],
     ) -> Result<(), VaultFailError>;
 }
-
-impl<D: Vault + Send + Sync + 'static> DynVault for D {
-    fn random(&mut self, data: &mut [u8]) -> Result<(), VaultFailError> {
-        Vault::random(self, data)
-    }
-
-    fn sha256(&self, data: &[u8]) -> Result<[u8; 32], VaultFailError> {
-        Vault::sha256(self, data)
-    }
-
-    fn secret_generate(
-        &mut self,
-        attributes: SecretKeyAttributes,
-    ) -> Result<Box<dyn Secret>, VaultFailError> {
-        Vault::secret_generate(self, attributes)
-    }
-
-    fn secret_import(
-        &mut self,
-        secret: &SecretKey,
-        attributes: SecretKeyAttributes,
-    ) -> Result<Box<dyn Secret>, VaultFailError> {
-        Vault::secret_import(self, secret, attributes)
-    }
-
-    fn secret_export(&mut self, context: &Box<dyn Secret>) -> Result<SecretKey, VaultFailError> {
-        Vault::secret_export(self, context)
-    }
-
-    fn secret_attributes_get(
-        &mut self,
-        context: &Box<dyn Secret>,
-    ) -> Result<SecretKeyAttributes, VaultFailError> {
-        Vault::secret_attributes_get(self, context)
-    }
-
-    fn secret_public_key_get(
-        &mut self,
-        context: &Box<dyn Secret>,
-    ) -> Result<PublicKey, VaultFailError> {
-        Vault::secret_public_key_get(self, context)
-    }
-
-    fn secret_destroy(&mut self, context: Box<dyn Secret>) -> Result<(), VaultFailError> {
-        Vault::secret_destroy(self, context)
-    }
-
-    fn ec_diffie_hellman(
-        &mut self,
-        context: &Box<dyn Secret>,
-        peer_public_key: PublicKey,
-    ) -> Result<Box<dyn Secret>, VaultFailError> {
-        Vault::ec_diffie_hellman(self, context, peer_public_key)
-    }
-
-    fn ec_diffie_hellman_hkdf_sha256(
-        &mut self,
-        context: &Box<dyn Secret>,
-        peer_public_key: PublicKey,
-        salt: &Box<dyn Secret>,
-        info: &[u8],
-        output_attributes: Vec<SecretKeyAttributes>,
-    ) -> Result<Vec<Box<dyn Secret>>, VaultFailError> {
-        Vault::ec_diffie_hellman_hkdf_sha256(
-            self,
-            context,
-            peer_public_key,
-            salt,
-            info,
-            output_attributes,
-        )
-    }
-
-    fn hkdf_sha256(
-        &mut self,
-        salt: &Box<dyn Secret>,
-        info: &[u8],
-        ikm: Option<&Box<dyn Secret>>,
-        output_attributes: Vec<SecretKeyAttributes>,
-    ) -> Result<Vec<Box<dyn Secret>>, VaultFailError> {
-        Vault::hkdf_sha256(self, salt, info, ikm, output_attributes)
-    }
-
-    fn aead_aes_gcm_encrypt(
-        &mut self,
-        context: &Box<dyn Secret>,
-        plaintext: &[u8],
-        nonce: &[u8],
-        aad: &[u8],
-    ) -> Result<Vec<u8>, VaultFailError> {
-        Vault::aead_aes_gcm_encrypt(self, context, plaintext, nonce, aad)
-    }
-
-    fn aead_aes_gcm_decrypt(
-        &mut self,
-        context: &Box<dyn Secret>,
-        cipher_text: &[u8],
-        nonce: &[u8],
-        aad: &[u8],
-    ) -> Result<Vec<u8>, VaultFailError> {
-        Vault::aead_aes_gcm_decrypt(self, context, cipher_text, nonce, aad)
-    }
-
-    fn deinit(&mut self) {
-        Vault::deinit(self)
-    }
-
-    fn sign(
-        &mut self,
-        secret_key: &Box<dyn Secret>,
-        data: &[u8],
-    ) -> Result<[u8; 64], VaultFailError> {
-        Vault::sign(self, secret_key, data)
-    }
-
-    fn verify(
-        &mut self,
-        signature: [u8; 64],
-        public_key: PublicKey,
-        data: &[u8],
-    ) -> Result<(), VaultFailError> {
-        Vault::verify(self, signature, public_key, data)
-    }
-}
